MS-939 View pulse animation helper extracted for general use
